Clean up schema usage
