+++ tests/conftest.py
@@ -3,6 +3,16 @@ from typing import Generator
 import pytest as pytest
 from chispa.dataframe_comparer import assert_df_equality
 from pyspark.sql import DataFrame, SparkSession
+from pyspark.sql.types import (
+    ArrayType,
+    BooleanType,
+    FloatType,
+    IntegerType,
+    LongType,
+    StringType,
+    StructField,
+    StructType,
+)
 
 
 @pytest.fixture(scope="session")
@@ -22,7 +32,7 @@ def spark() -> Generator:
 
 
 def assert_data_frames_equal(left: DataFrame, right: DataFrame) -> None:
-    assert_df_equality(left, right, ignore_row_order=True, ignore_schema=True)
+    assert_df_equality(left, right, ignore_row_order=True, ignore_nullable=True)
 
 
 def _drop_database_cascade(spark: SparkSession, db: str) -> None:
@@ -34,3 +44,82 @@ def _drop_database_cascade(spark: SparkSession, db: str) -> None:
     tables = spark.sql(f"SHOW TABLES IN {db}").collect()
     [spark.sql(f"DROP TABLE {db}.{t.tableName}") for t in tables]
     spark.sql(f"DROP DATABASE {db}")
+
+
+class Schema:
+    RAW = StructType(
+        [
+            StructField("movie_id", LongType()),
+            StructField("user_id", LongType()),
+            StructField("rating", FloatType()),
+            StructField("timestamp", LongType()),
+            StructField("original_title", StringType()),
+            StructField("original_language", StringType()),
+            StructField("budget", LongType()),
+            StructField("adult", BooleanType()),
+            StructField(
+                "genres",
+                ArrayType(
+                    StructType(
+                        [
+                            StructField("id", LongType()),
+                            StructField("name", StringType()),
+                        ]
+                    )
+                ),
+                nullable=False,
+            ),
+        ]
+    )
+
+    STANDARDIZED = StructType(
+        [
+            StructField("movie_id", LongType()),
+            StructField("user_id", LongType()),
+            StructField("rating", FloatType()),
+            StructField("timestamp", LongType()),
+            StructField("original_title", StringType()),
+            StructField("original_language", StringType()),
+            StructField("budget", LongType()),
+            StructField("adult", BooleanType()),
+            StructField(
+                "genres",
+                ArrayType(
+                    StructType(
+                        [
+                            StructField("id", LongType()),
+                            StructField("name", StringType()),
+                        ]
+                    )
+                ),
+            ),
+            StructField("run_date", IntegerType()),
+        ]
+    )
+
+    CURATED = StructType(
+        [
+            StructField("movie_id", LongType()),
+            StructField("user_id", LongType()),
+            StructField("rating", FloatType()),
+            StructField("rating_class", StringType()),
+            StructField("timestamp", LongType()),
+            StructField("original_title", StringType()),
+            StructField("original_language", StringType()),
+            StructField("budget", LongType()),
+            StructField("is_adult", BooleanType()),
+            StructField("is_multigenre", BooleanType()),
+            StructField(
+                "genres",
+                ArrayType(
+                    StructType(
+                        [
+                            StructField("id", LongType()),
+                            StructField("name", StringType()),
+                        ]
+                    )
+                ),
+            ),
+            StructField("run_date", IntegerType()),
+        ]
+    )
